requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
playwright==1.40.0

# RAG and Vector Storage
langchain==0.0.350
langchain-community==0.0.10
chromadb==0.4.18
faiss-cpu==1.12.0
sentence-transformers==2.2.2
tiktoken==0.5.1

# Google AI and APIs
google-generativeai==0.3.2
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Database and Storage
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# File Processing
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
celery==5.3.4
pydantic-settings==2.0.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Database setup
alembic==1.12.1
